+++ helper/soundcapsule/__init__.py
@@ -11,4 +11,4 @@ __all__ = [
     "FLPUnsupportedError",
 ]
 
-__version__ = "0.1.0"
+__version__ = "0.2.0"